Mission Demolition Camera and Castles
Added some stuff to make the camera more reliable,, made it so you can fire multiple times, added clouds for a sense of space and scale, and I created a castle with physics that can be destroyed and has a goal. Will fix issues with projectiles/structures not sleeping next time. Page 45